Added template set feature
Allow bconf to support several templates in its datadir
installation prefix, to support different default use-cases.
Added a "host-only" template set, which only compiles for the host machine
and a "build-aware" template set, which adds a rule to compile a binary
for the build machine, if a custom-utility is required at build time.